refactor TxValueDisplay
