+++ common/v2/features/SendAssets/components/index.ts
@@ -1,5 +1,5 @@
 export { default as ConfirmTransaction } from './ConfirmTransaction';
-export { default as SendAssetsForm } from './TransactionFormData';
+export { default as SendAssetsForm } from './SendAssetsForm';
 export { default as TransactionReceipt } from './TransactionReceipt';
 export { default as SignTransaction } from './SignTransaction';
 export * from './validators';
